pyproject.toml
[project]
name = "graduation-thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.0.3",
    "langchain-dev-utils>=1.1.7",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.13",
    "mineru-mcp>=1.0.0",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
]

